model/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-openai
langgraph
opencv-python
numpy
python-multipart
google-auth-oauthlib
google-api-python-client
python-multipart